Apply all setting updates and pull them from parameter store
