+++ prdeploy-api/scripts/GraphQlScaffold.ps1
@@ -1,43 +1,19 @@
-param (
-    [ValidateSet('all', 'update', 'build')]
-    [string]
-    $command =  'all'
-)
-
-. "$PSScriptRoot/_Supporting.ps1"
-
 $repoRoot = [System.IO.Path]::GetFullPath("$PSScriptRoot/..")
+$testDirectory = "$repoRoot/test/PrDeploy.Api.Tests"
 
-if (($command -eq 'all') -or ($command -eq 'update')) {
+Push-Location "$testDirectory/Client"
 
-    # SourceRef: https://chillicream.com/docs/strawberryshake/v12/get-started/console
-    Write-Header "Updating GraphQL schema files"
-
-    Push-Location "$repoRoot/test/PrDeploy.Api.Tests/Client"
-
-    dotnet graphql update
-
-    if ($LASTEXITCODE -ne 0) {
-        Write-Host "`nMake sure the GraphQL server is started." -ForegroundColor Yellow
-        Pop-Location
-
-        Receive-Job $graphQlApi | Out-Null
-        Stop-Job $graphQlApi
-        exit -1
-    }
-
-    Pop-Location
+Write-Host "Updating schema in Test project."
+dotnet graphql update
+if ($LASTEXITCODE -ne 0) {
+  Write-Host "Make sure you are running the API first." -ForegroundColor Yellow
+  Pop-Location
+  exit 1
 }
 
-if (($command -eq 'all') -or ($command -eq 'build')) {
+Write-Host "Cleaning generated GraphQL classes."
+Remove-Item "$testDirectory/obj/Debug/net6.0/berry" -Recurse -Force
 
-    Write-Header "Building Tests to generate classes"
-
-    Push-Location "$repoRoot/test/PrDeploy.Api.Tests"
-
-    dotnet build .\PrDeploy.Api.Tests.csproj
-
-    Pop-Location
-}
+Pop-Location
 
 Write-Host "`nGraphQL scaffold complete." -ForegroundColor Green
