+++ prdeploy-api/src/PrDeploy.Api.Business/Options/AwsOptions.cs
@@ -1,9 +1,9 @@
 namespace PrDeploy.Api.Business.Options;
 public class AwsOptions
 {
-    public string AccessKeyId { get; set; }
+    public string AccessKeyId { get; set; } = string.Empty;
 
-    public string SecretAccessKey { get; set; }
+    public string SecretAccessKey { get; set; } = string.Empty;
 
-    public string Region { get; set; }
+    public string Region { get; set; } = string.Empty;
 }
