+++ prdeploy-api/src/PrDeploy.Api.Models/Settings/Compare/JiraSettingsCompare.cs
@@ -2,9 +2,9 @@
 {
     public class JiraSettingsCompare
     {
-        public OwnerRepoValue<bool?> AddIssuesEnabled { get; set; }
-        public OwnerRepoValue<string> Host { get; set; }
-        public OwnerRepoValue<string> Username { get; set; }
-        public OwnerRepoValue<string> Password { get; set; }
+        public OwnerRepoValue<bool?> AddIssuesEnabled { get; set; } = null!;
+        public OwnerRepoValue<string> Host { get; set; } = null!;
+        public OwnerRepoValue<string> Username { get; set; } = null!;
+        public OwnerRepoValue<string> Password { get; set; } = null!;
     }
 }
