+++ prdeploy-api/src/PrDeploy.Api.Models/Settings/ServiceSettings.cs
@@ -1,6 +1,6 @@
 namespace PrDeploy.Api.Models.Settings;
 public class ServiceSettings
 {
-    public string? Name { get; set; }
+    public string Name { get; set; } = string.Empty;
     public string? Path { get; set; }
 }
