+++ prdeploy-webhooks/.env.sample
@@ -5,11 +5,9 @@ AWS_REGION="us-west-2"          # TODO GBJ: Use local AWS role instead.
 AWS_ACCESS_KEY_ID=""
 AWS_SECRET_ACCESS_KEY=""
 GITHUB_TOKEN=""
-SETTINGS_BRANCH=""
 ENABLE_OCTOKIT_LOGGING="true"
 ENABLE_CONSOLE_LOGGING="false"
 TEMPLATE_PATH="./src/templates"
-DEFAULT_SETTINGS_FILE="./src/settings/.default-prdeploy.yaml"
 
 # Test settings only.
 LOG_LEVEL="debug"
