+++ prdeploy-webhooks/src/webhook-flow.ts
@@ -1,16 +1,16 @@
 import { Octokit } from '@octokit/rest';
 import { DependencyContainer, container } from 'tsyringe';
-import { LogService, RepoSettingsService } from './services';
+import { LogService, DeploySettingsService } from './services';
 import { Repository } from '@octokit/webhooks-types';
 import { REPOSITORY } from './injection-tokens';
-import { PullRequest, RepoSettings } from './models';
+import { PullRequest, DeploySettings } from './models';
 import JiraApi from 'jira-client';
 
 export const webhookFlow = async (
   octokit: Octokit | any,
   repository: Repository | any,
   pullRequest: PullRequest | any,
-  action: (scope: DependencyContainer, settings: RepoSettings, log: LogService) => Promise<void>
+  action: (scope: DependencyContainer, settings: DeploySettings, log: LogService) => Promise<void>
 ): Promise<void> => {
   const childScope = container.createChildContainer();
   childScope.register(Octokit, { useValue: octokit });
@@ -20,9 +20,9 @@ export const webhookFlow = async (
     repository,
     pullRequest
   });
-  const settingsService = childScope.resolve(RepoSettingsService);
+  const settingsService = childScope.resolve(DeploySettingsService);
   const settings = await settingsService.get();
-  childScope.register(RepoSettings, { useValue: settings });
+  childScope.register(DeploySettings, { useValue: settings });
   childScope.register(JiraApi, {
     useFactory: () => {
       return new JiraApi({
