--- prdeploy-webhooks/test/settings/.test-default-prdeploy.yaml
@@ -1,58 +1,0 @@
-# All settings showing defaults, only override what needs to change.
-
-deployWorkflow: combined-deploy.yaml
-environments:
-  - name: dev
-    queue: 'DEV_PR_QUEUE'
-    url: 'https://dev.myorg.com'
-  - name: dev2
-    queue: 'DEV_PR_QUEUE'
-    url: 'https://dev2.myorg.com'
-  - name: dev3
-    queue: 'DEV_PR_QUEUE'
-    url: 'https://dev3.myorg.com'
-  - name: stage
-    queue: 'STAGE_PR_QUEUE'
-    url: 'https://stage.myorg.com'
-    requireBranchUpToDate: true
-    automationTest:
-      enabled: true
-      workflow: test-automation-run.yml
-      inputs:
-        environment: ${environment}
-        pull_number: ${pull_number}
-        run_name: ${run_name}
-  - name: prod
-    queue: 'PROD_PR_QUEUE'
-    url: 'https://www.myorg.com'
-    requireBranchUpToDate: true
-    requireApproval: true
-    excludeFromRollback:
-      - myorg-db
-    automationTest:
-      enabled: false
-      workflow: test-automation-run.yml
-      inputs:
-        environment: ${environment}
-        pull_number: ${pull_number}
-        run_name: ${run_name}
-services:
-  - name: myorg-app-main
-  - name: myorg-api-service
-  - name: myorg-db
-defaultEnvironment: dev
-releaseEnvironment: prod
-settingsBranch: 'main'
-builds:
-  checkPattern: 'Build helm chart'
-  workflowPattern: '/([^/]+?)-build.ya?ml'
-jira:
-  addIssuesEnabled: true
-slack:
-  notificationsEnabled: true
-badge:
-  statusColors:
-    error: 'cc2900'
-    warn: 'cccc00'
-    success: '29a329'
-    info: '0080ff'
